use new projection query function in libpano13, projections are now queried dynamically

+++ src/nona_gui/nona_gui.cpp
@@ -241,7 +241,7 @@ bool nonaApp::OnInit()
             wxLogError(_("Wrong number of images specified on command line"));
             return false;
         }
-        for (int i = 0; i < pano.getNrOfImages(); i++) {
+        for (size_t i = 0; i < pano.getNrOfImages(); i++) {
             pano.setImageFilename(i, (const char *)parser.GetParam(i+1).mb_str());
         }
     }
